+++ R/create.clone.genome.distribution.densityplot.R
@@ -12,7 +12,7 @@ create.clone.genome.distribution.densityplot <- function(
         data = density.df,
         groups = density.df$clone.id,
         xlab.label = 'Chromosome',
-        ylab.label = 'Number of SNVs',
+        ylab.label = 'SNV Density',
         xlimits = c(0, sum(chr.info$length)),
         xaxis.lab = chr.info$chr,
         xat = chr.info$xat,
@@ -26,7 +26,12 @@ create.clone.genome.distribution.densityplot <- function(
     }
 
 calculate.density.and.scale <- function(cluster.df) {
-    density <- density(x = cluster.df$genome.pos, bw = 'nrd', adjust = 0.05, na.rm = TRUE);
+    # density should be generated using unque SNV count
+    density <- density(
+        x = cluster.df$genome.pos,
+        bw = 'nrd',
+        adjust = 0.05, # set to 1E9/3E9 to get density per megabase
+        na.rm = TRUE);
     density.df <- as.data.frame(density[c('x','y')]);
     density.df$clone.id <- unique(cluster.df$clone.id);
     density.df$count <- nrow(cluster.df) / sum(density.df$y) * density.df$y;